+++ src/functions/numeric/reciprocal.js
@@ -1,0 +1,28 @@
+// Include
+const LimeFunction = require('../../structs/function');
+
+/* ------------------------ division ------------------------ */
+
+// Reciprocal function class
+class LimeFunctionReciprocal extends LimeFunction {
+  // Constructor
+  constructor(lime, mode) {
+    // Super from function class
+    super(lime, { name: 'reciprocal', mode });
+
+    // Operations
+    this.operations.r = [
+      'er(arg{expr})',
+    ];
+
+    // Algorithms
+    this.algorithms.set('r(arg{expr})', (step) => {
+      step.rus(this.lime.direct([1, '/', step.right.places[0]]));
+    });
+  }
+}
+
+/* ------------------------ division ------------------------ */
+
+// Export
+module.exports = LimeFunctionReciprocal;