+++ src/system/oplist.js
@@ -41,8 +41,29 @@ const Oplist = {
     // Nested
     'n()': () => true,
 
+    // [Right unary] Argument{}
+    'r(arg{})': (step) => step.rpi('argument') && step.right.isEmpty,
+
     // [Right unary] Argument{Expression}
     'r(arg{expr})': (step) => step.rpi('argument') && step.right.length === 1 && step.ci('expression')(step.right.places[0]),
+
+    // [Right unary] Argument{Expression[2]}
+    'r(arg{expr[2]})': (step) => step.rpi('argument') && step.right.length === 2 && step.ci('expression')(step.right.places[0], step.right.places[1]),
+
+    // [Right unary] Argument{Expression[Any]}
+    'r(arg{expr[@]})': (step) => step.rpi('argument') && step.ci('expression')(...step.right.places),
+
+    // [Right unary] Argument{Integer}
+    'r(arg{int})': (step) => step.rpi('argument') && step.right.length === 1 && step.ci('integer')(step.right.places[0]),
+
+    // [Right unary] Argument{Integer[2]}
+    'r(arg{int[2]})': (step) => step.rpi('argument') && step.right.length === 2 && step.ci('integer')(step.right.places[0], step.right.places[1]),
+
+    // [Right unary] Argument{Integer[Any]}
+    'r(arg{int[@]})': (step) => step.rpi('argument') && step.ci('integer')(...step.right.places),
+
+    // [Right unary] Argument{Rational}
+    'r(arg{rat})': (step) => step.rpi('argument') && step.right.length === 1 && step.ci('rational')(step.right.places[0]),
 
     // [Right unary] Boolean
     'r(bool)': (step) => step.rpi('boolean'),
@@ -157,7 +178,7 @@ const Oplist = {
 
     // Convert left parameter from argument length 1 to expression
     'l(arg{expr}->expr)': (step) => {
-      step.lps(step.left.simplify());
+      step.lps(step.left.finalize());
     },
 
     // Convert left parameter from boolean to integer
@@ -182,7 +203,7 @@ const Oplist = {
 
     // Convert right parameter from argument length 1 to expression
     'r(arg{expr}->expr)': (step) => {
-      step.rps(step.right.simplify());
+      step.rps(step.right.finalize());
     },
 
     // Convert right parameter from boolean to integer